feat: лайки и дизлайки доделаны, состояния обновляются сразу
