+++ src/components/SavedMovies/SavedMovies.jsx
@@ -4,7 +4,10 @@ import MoviesCardList from "../MoviesCardList/MoviesCardList";
 import Footer from "../Footer/Footer";
 import useMoviesSearch from "../../utils/useMoviesSearch";
 import Preloader from "../Preloader/Preloader";
-export default function SavedMovies({ savedMovies, errorFetchSavedMovies }) {
+import { useEffect } from 'react';
+
+export default function SavedMovies({ savedMovies, errorFetchSavedMovies, handleLikeClick }) {
+
   const {
     searchedMovies,
     isSearching,
@@ -15,6 +18,10 @@ export default function SavedMovies({ savedMovies, errorFetchSavedMovies }) {
     handleSearch,
     handleToggleShortFilmOnly,
   } = useMoviesSearch(savedMovies);
+
+  useEffect(() => {
+    handleSearch(searchQuery, shortFilmOnly);
+  }, [savedMovies]);
 
   return (
     <main className={'movies'}>
@@ -32,7 +39,7 @@ export default function SavedMovies({ savedMovies, errorFetchSavedMovies }) {
         <>
           {searchedMovies.length > 0 ? (
             <>
-              <MoviesCardList movies={searchedMovies} />
+              <MoviesCardList movies={searchedMovies} handleLikeClick={handleLikeClick} />
             </>
           ) : (
             <p className={'error-message'}>
